dispatcher/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
pymongo==4.8.0
python-jose==3.3.0
pytest==8.3.3
